+++ arma/src/esm/src/endpoints/increment_territory_payment_counter.rs
@@ -12,9 +12,9 @@ pub fn increment_territory_payment_counter(
     );
 
     // Convert the database Ids from "['1','2']" to [1,2]
-    let database_ids: Vec<usize> =
+    let database_ids: Vec<u64> =
         match Parser::from_arma::<Vec<String>>(&database_ids) {
-            Ok(ids) => ids.iter().filter_map(|i| i.parse::<usize>().ok()).collect(),
+            Ok(ids) => ids.iter().filter_map(|i| i.parse::<u64>().ok()).collect(),
             Err(e) => return Err(e),
         };
